Add basic login form; Move css styling to index.css
Why?
Tl;dr This is temporary and should be changed.
Each React container may override previous styling.
It is dependent on source loading ordering. I moved
styling to index.css for now in order to avoid this
behaviour and still be able to use Bootstrap (css
modules won't probably allow me to do it in easy way).
I will look up better way to handle it in the future.